requirements.txt
fastmcp
httpx
python-dotenv
pydantic
opentelemetry-api
opentelemetry-sdk
fastapi
uvicorn